+++ src/main/java/com/Backend/ToothDay/visit/repository/VisitRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VisitRepository extends JpaRepository<Visit, Integer> {
+public interface VisitRepository extends JpaRepository<Visit, Long> {
 
 }
